estudo sobre abstractfactory

+++ DesignPatternOO/app/src/main/java/br/com/xplorer/designpatternoo/entities/Baguette.java
@@ -1,5 +1,7 @@
 package br.com.xplorer.designpatternoo.entities;
 
+import br.com.xplorer.designpatternoo.entities.abs.Bread;
+
 /**
  * Created by r028367 on 22/03/2018.
  */
@@ -12,7 +14,7 @@ public class Baguette implements Bread {
     }
 
     @Override
-    public long getCalories() {
+    public double getCalories() {
         return 0;
     }
 }
