+++ DesignPatternOO/app/src/main/java/br/com/xplorer/designpatternoo/entities/Brioche.java
@@ -1,6 +1,8 @@
 package br.com.xplorer.designpatternoo.entities;
 
 
+import br.com.xplorer.designpatternoo.entities.abs.Bread;
+
 /**
  * Created by r028367 on 23/03/2018.
  */
@@ -13,7 +15,7 @@ public class Brioche implements Bread {
     }
 
     @Override
-    public long getCalories() {
+    public double getCalories() {
         return 0;
     }
 }
